requirements.txt
Django~=4.0.2
django-mapbox-location-field~=2.0.0
Pillow~=9.0.1